web/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
numpy==1.20.2
streamlit==0.80.0
cassava-classifier==0.0.2